Put functions in suitable modules
